Refactor report details form wizards and remove unnecessary cancel link
