+++ server/routes/index.ts
@@ -6,11 +6,11 @@ import type { Services } from '../services'
 import { PrisonApi } from '../data/prisonApi'
 import makeDebugRoutes from './debug'
 import makeDownloadConfigRouter from './downloadReportConfig'
-import { createReportRouter } from './reports/createReport'
+import { createReportRouter } from './reports/details/createReport'
+import { updateDetailsRouter } from './reports/details/updateReportDetails'
 import { prisonerInvolvementRouter } from './reports/prisoners'
 import { staffInvolvementRouter } from './reports/staff'
 import { questionsRouter } from './reports/questions'
-import { updateDetailsRouter } from './reports/updateReportDetails'
 import { viewReportRouter } from './reports/viewReport'
 import dashboard from './dashboard/dashboard'
 import { dprRouter } from './dpr'
