+++ server/routes/reports/details/detailsFields.ts
