+++ server/routes/reports/details/updateReportDetails.ts
@@ -1,14 +1,14 @@
 import express from 'express'
 import FormWizard from 'hmpo-form-wizard'
 
-import logger from '../../../logger'
-import format from '../../utils/format'
-import type { ReportBasic } from '../../data/incidentReportingApi'
+import logger from '../../../../logger'
+import format from '../../../utils/format'
+import type { ReportBasic } from '../../../data/incidentReportingApi'
+import { logoutIf } from '../../../middleware/permissions'
+import { populateReport } from '../../../middleware/populateReport'
+import { cannotEditReport } from '../permissions'
 import { BaseDetailsController } from './detailsController'
 import { type DetailsValues, type DetailsFieldNames, detailsFields, detailsFieldNames } from './detailsFields'
-import { logoutIf } from '../../middleware/permissions'
-import { populateReport } from '../../middleware/populateReport'
-import { cannotEditReport } from './permissions'
 
 class DetailsController extends BaseDetailsController<DetailsValues> {
   // TODO: wizard namespace identifier is shared. consider generating it per request somehow?
@@ -34,12 +34,12 @@ class DetailsController extends BaseDetailsController<DetailsValues> {
     req.sessionModel.set('_incidentTime-minutes', minutes)
     req.sessionModel.set('description', report.description)
 
-    res.locals.cancelUrl = `/reports/${reportId}`
+    res.locals.backLinkUrl = `/reports/${reportId}`
     next()
   }
 
   getBackLink(_req: FormWizard.Request<DetailsValues, DetailsFieldNames>, res: express.Response): string {
-    return res.locals.cancelUrl
+    return res.locals.backLinkUrl
   }
 
   async successHandler(
@@ -76,7 +76,7 @@ class DetailsController extends BaseDetailsController<DetailsValues> {
 
   getNextStep(_req: FormWizard.Request<DetailsValues, DetailsFieldNames>, res: express.Response): string {
     // TODO: does this page have 2 save buttons? where do they both lead?
-    return res.locals.cancelUrl
+    return res.locals.backLinkUrl
   }
 }
